requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
certifi==2019.6.16
cffi==1.12.3
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
dj-database-url==0.5.0
Django==2.2.2
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.3.4
mysqlclient==1.4.2
parsel==1.5.1
pipenv==2018.11.26
prettytable==0.7.2
psycopg2==2.8.3
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylibscrypt==1.8.0
pyOpenSSL==19.0.0
pytz==2019.1
queuelib==1.5.0
Scrapy==1.6.0
selenium==3.141.0
service-identity==18.1.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
w3lib==1.20.0
whitenoise==4.1.2
zope.interface==4.6.0
